requirements.txt
black==23.9.1
blinker==1.6.2
click==8.1.7
exceptiongroup==1.1.3
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytest==7.4.2
SQLAlchemy==2.0.20
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.7
